heaps of changes
The main CSS file styles.css is now importing all styles from the other 3 CSS files. I had to add unique classes so there wouldn't be any issues between the pages and styling. In the future it would be cool to use similar styles in the main and unique styles would be in a separate file as well as the media queries.